requirement.txt
streamlit
networkx
openai
python-dotenv
numpy
pydantic